[IMP] product,_*: packaging button in catalog
_* : sale, purchase
Prior this commit
- Packaging button in catalog was only availabe in purchase

Post this commit
- Moved the packaging button logic to product
- Now we can add packaging button to a model inheriting catalog
  by passing `packaging` to `_get_product_catalog_order_data`
  and `_get_product_catalog_lines_data` and handling the return
  in `_update_order_line_info`

task-4200918
