+++ addons/product/static/src/product_catalog/order_line/order_line.js
@@ -9,6 +9,7 @@ export class ProductCatalogOrderLine extends Component {
         quantity: Number,
         price: Number,
         productType: String,
+        packaging: { type: Object, optional: true },
         readOnly: { type: Boolean, optional: true },
         warning: { type: String, optional: true},
     };
@@ -19,6 +20,10 @@ export class ProductCatalogOrderLine extends Component {
      */
     _onFocus(ev) {
         ev.target.select();
+    }
+
+    addPackagingQty() {
+        this.env.updatePackagingQuantity(this.props.packaging);
     }
 
     //--------------------------------------------------------------------------
