+++ addons/purchase/static/src/product_catalog/purchase_order_line/purchase_order_line.js
@@ -14,8 +14,4 @@ export class ProductCatalogPurchaseOrderLine extends ProductCatalogOrderLine {
     get highlightUoM() {
         return true;
     }
-
-    addPackagingQty() {
-        this.env.updatePackagingQuantity(this.props.packaging);
-    }
 }
