+++ addons/purchase/static/src/product_catalog/purchase_order_line/purchase_order_line.xml
@@ -15,12 +15,5 @@
                 </span>
             </span>
         </xpath>
-        <xpath expr="//div[hasclass('o_product_catalog_quantity')]" position="inside">
-            <button t-if="props.packaging"
-                    class="o_product_catalog_package_quantity btn btn-primary border"
-                    t-on-click.stop="addPackagingQty">
-                <i class="fa fa-cube"/>
-            </button>
-        </xpath>
     </t>
 </templates>
